BK-749 Moved epub related utils to separate module
